separate compilation units for passes
