+++ samples/verona/passes/assignlhs.cc
@@ -1,0 +1,72 @@
+// Copyright Microsoft and Project Verona Contributors.
+// SPDX-License-Identifier: MIT
+#include "lang.h"
+
+namespace verona
+{
+  auto on_lhs(auto pattern)
+  {
+    return (In(Assign) * (pattern * ++T(Expr))) / (In(TupleLHS) * pattern);
+  }
+
+  PassDef assignlhs()
+  {
+    return {
+      // Turn a Tuple on the LHS of an assignment into a TupleLHS.
+      on_lhs(T(Expr) << T(Tuple)[Lhs]) >>
+        [](Match& _) { return Expr << (TupleLHS << *_[Lhs]); },
+
+      on_lhs(T(Expr) << (T(TypeAssert) << (T(Tuple)[Lhs] * T(Type)[Type]))) >>
+        [](Match& _) {
+          return Expr << (TypeAssert << (TupleLHS << *_[Lhs]) << _(Type));
+        },
+
+      // Turn a Call on the LHS of an assignment into a CallLHS.
+      on_lhs(T(Expr) << T(Call)[Lhs]) >>
+        [](Match& _) { return Expr << (CallLHS << *_[Lhs]); },
+
+      on_lhs(T(Expr) << (T(TypeAssert) << (T(Call)[Lhs] * T(Type)[Type]))) >>
+        [](Match& _) {
+          return Expr << (TypeAssert << (CallLHS << *_[Lhs]) << _(Type));
+        },
+
+      // Turn a RefVar on the LHS of an assignment into a RefVarLHS.
+      on_lhs(T(Expr) << T(RefVar)[Lhs]) >>
+        [](Match& _) { return Expr << (RefVarLHS << *_[Lhs]); },
+
+      on_lhs(T(Expr) << (T(TypeAssert) << (T(RefVar)[Lhs] * T(Type)[Type]))) >>
+        [](Match& _) {
+          return Expr << (TypeAssert << (RefVarLHS << *_[Lhs]) << _(Type));
+        },
+
+      In(Expr) * T(Ref)[Ref] >>
+        [](Match& _) {
+          return err(_[Ref], "must use `ref` in front of a variable or call");
+        },
+
+      In(Expr) * T(Try)[Try] >>
+        [](Match& _) {
+          return err(_[Try], "must use `try` in front of a call or lambda");
+        },
+
+      T(Expr)[Expr] << (Any * Any * Any++) >>
+        [](Match& _) {
+          return err(_[Expr], "adjacency on this expression isn't meaningful");
+        },
+
+      In(TupleLHS) * T(TupleFlatten) >>
+        [](Match& _) {
+          return err(
+            _[TupleFlatten],
+            "can't flatten a tuple on the left-hand side of an assignment");
+        },
+
+      In(Expr) * T(Expr)[Expr] >>
+        [](Match& _) {
+          return err(
+            _[Expr],
+            "well-formedness allows this but it can't occur on written code");
+        },
+    };
+  }
+}
