+++ samples/verona/passes/defbeforeuse.cc
@@ -1,0 +1,25 @@
+// Copyright Microsoft and Project Verona Contributors.
+// SPDX-License-Identifier: MIT
+#include "lang.h"
+
+namespace verona
+{
+  PassDef defbeforeuse()
+  {
+    return {
+      dir::topdown | dir::once,
+      {
+        T(RefLet) << T(Ident)[Id] >> ([](Match& _) -> Node {
+          auto id = _(Id);
+          auto defs = id->lookup();
+
+          if (
+            (defs.size() == 1) &&
+            ((defs.front()->type() == Param) || defs.front()->precedes(id)))
+            return NoChange;
+
+          return err(_[Id], "use of uninitialized identifier");
+        }),
+      }};
+  }
+}
