+++ samples/verona/passes/reference.cc
@@ -1,0 +1,88 @@
+// Copyright Microsoft and Project Verona Contributors.
+// SPDX-License-Identifier: MIT
+#include "lang.h"
+
+#include "lookup.h"
+
+namespace verona
+{
+  PassDef reference()
+  {
+    return {
+      // LLVM literal.
+      In(Expr) * T(LLVM)[LLVM] * T(Ident)[Lhs] * T(Ident)++[Rhs] >>
+        [](Match& _) {
+          auto llvm = _(LLVM);
+          auto rhs = _[Rhs];
+          auto s = std::string()
+                     .append(llvm->location().view())
+                     .append(" %")
+                     .append(_(Lhs)->location().view());
+
+          for (auto& i = rhs.first; i != rhs.second; ++i)
+            s.append(", %").append((*i)->location().view());
+
+          return LLVM ^ s;
+        },
+
+      In(Expr) * T(LLVM)[Lhs] * T(LLVM)[Rhs] >>
+        [](Match& _) {
+          return LLVM ^
+            std::string()
+              .append(_(Lhs)->location().view())
+              .append(" ")
+              .append(_(Rhs)->location().view());
+        },
+
+      // Dot notation. Use `Id` as a selector, even if it's in scope.
+      In(Expr) * T(Dot) * Name[Id] * ~T(TypeArgs)[TypeArgs] >>
+        [](Match& _) {
+          return Seq << Dot << (Selector << _[Id] << (_(TypeArgs) || TypeArgs));
+        },
+
+      // Local reference.
+      In(Expr) * T(Ident)[Id]([](auto& n) { return lookup(n, {Var}); }) >>
+        [](Match& _) { return RefVar << _(Id); },
+
+      In(Expr) * T(Ident)[Id]([](auto& n) {
+        return lookup(n, {Let, Param});
+      }) >>
+        [](Match& _) { return RefLet << _(Id); },
+
+      // Unscoped type reference.
+      In(Expr) * T(Ident)[Id]([](auto& n) {
+        return lookup(n, {Class, TypeAlias, TypeParam});
+      }) * ~T(TypeArgs)[TypeArgs] >>
+        [](Match& _) {
+          return makename(DontCare, _(Id), (_(TypeArgs) || TypeArgs));
+        },
+
+      // Unscoped reference that isn't a local or a type. Treat it as a
+      // selector, even if it resolves to a Function.
+      In(Expr) * Name[Id] * ~T(TypeArgs)[TypeArgs] >>
+        [](Match& _) { return Selector << _(Id) << (_(TypeArgs) || TypeArgs); },
+
+      // Scoped lookup.
+      In(Expr) *
+          (TypeName[Lhs] * T(DoubleColon) * Name[Id] *
+           ~T(TypeArgs)[TypeArgs])[Type] >>
+        [](Match& _) {
+          return makename(_(Lhs), _(Id), (_(TypeArgs) || TypeArgs), true);
+        },
+
+      In(Expr) * T(DoubleColon) >>
+        [](Match& _) { return err(_[DoubleColon], "expected a scoped name"); },
+
+      // Create sugar, with no arguments.
+      In(Expr) * TypeName[Lhs] * ~T(TypeArgs)[TypeArgs] >>
+        [](Match& _) {
+          return FunctionName << _(Lhs) << (Ident ^ create)
+                              << (_(TypeArgs) || TypeArgs);
+        },
+
+      // Lone TypeArgs are typeargs on apply.
+      In(Expr) * T(TypeArgs)[TypeArgs] >>
+        [](Match& _) { return Seq << Dot << apply(_(TypeArgs)); },
+    };
+  }
+}
