+++ samples/verona/passes/typeflat.cc
@@ -1,0 +1,39 @@
+// Copyright Microsoft and Project Verona Contributors.
+// SPDX-License-Identifier: MIT
+#include "lang.h"
+
+namespace verona
+{
+  PassDef typeflat()
+  {
+    return {
+      // Flatten algebraic types.
+      In(TypeUnion) * T(TypeUnion)[Lhs] >>
+        [](Match& _) { return Seq << *_[Lhs]; },
+      In(TypeIsect) * T(TypeIsect)[Lhs] >>
+        [](Match& _) { return Seq << *_[Lhs]; },
+      In(TypeView) * T(TypeView)[Lhs] >>
+        [](Match& _) { return Seq << *_[Lhs]; },
+
+      // Tuples of arity 1 are scalar types.
+      T(TypeTuple) << (TypeElem[Op] * End) >> [](Match& _) { return _(Op); },
+
+      // Tuples of arity 0 are the unit type.
+      T(TypeTuple) << End >> [](Match&) { return unittype(); },
+
+      // Flatten Type nodes. The top level Type node won't go away.
+      TypeStruct * T(Type) << (TypeElem[Op] * End) >>
+        [](Match& _) { return _(Op); },
+
+      T(Type)[Type] << End >>
+        [](Match& _) {
+          return err(_[Type], "can't use an empty type assertion");
+        },
+
+      T(Type)[Type] << (Any * Any) >>
+        [](Match& _) {
+          return err(_[Type], "can't use adjacency to specify a type");
+        },
+    };
+  }
+}
