+++ samples/verona/passes/typeview.cc
@@ -1,0 +1,32 @@
+// Copyright Microsoft and Project Verona Contributors.
+// SPDX-License-Identifier: MIT
+#include "lang.h"
+
+namespace verona
+{
+  PassDef typeview()
+  {
+    return {
+      // Viewpoint adaptation binds more tightly than function types.
+      TypeStruct * TypeElem[Lhs] * T(Dot) * TypeElem[Rhs] >>
+        [](Match& _) {
+          return TypeView << (Type << _[Lhs]) << (Type << _[Rhs]);
+        },
+
+      // TypeList binds more tightly than function types.
+      TypeStruct * TypeElem[Lhs] * T(Ellipsis) >>
+        [](Match& _) { return TypeList << (Type << _[Lhs]); },
+
+      TypeStruct * T(DoubleColon)[DoubleColon] >>
+        [](Match& _) { return err(_[DoubleColon], "misplaced type scope"); },
+      TypeStruct * T(TypeArgs)[TypeArgs] >>
+        [](Match& _) {
+          return err(_[TypeArgs], "type arguments on their own are not a type");
+        },
+      TypeStruct * T(Dot)[Dot] >>
+        [](Match& _) { return err(_[Dot], "misplaced type viewpoint"); },
+      TypeStruct * T(Ellipsis)[Ellipsis] >>
+        [](Match& _) { return err(_[Ellipsis], "misplaced type list"); },
+    };
+  }
+}
